requirements.txt
discord.py
python-dotenv
flask
asyncpg